+++ packages/app/src/pages/_/ak/_id/_.route.vue
@@ -1,11 +1,12 @@
-<template lang="pug">
-  
+<template lang="pug">  
   ec-wrapper(hasXBtn hasSheet hasHeader hasDial v-bind="config" hasReload @reload="loadData")
+    //- Kopf
     template(#header)
       .head(style="padding: 2px 10px")
         v-switch(label="Alle Statusupdates anzeigen?" v-model="showAll")
         h2(v-font v-primary v-if="showAll") Alle Statusupdates
         h2(v-font v-primary v-else) Aktuelle Mitglieder
+    //- Main Content
     v-list(two-line)
       template(v-if="data.personen" v-for="(person, index) in data.personen")
         v-list-tile(v-if="!showAll && person.currentStatus!==0" :key="person.person.personID + '_c'" @click="$router.push({path: `/personen/${person.person.personID}/home`, query: {prev: $route.fullPath}})")
@@ -21,40 +22,7 @@
           v-list-tile-content
             v-list-tile-title {{person.person.vorname}} {{person.person.nachname}} ({{person.person.gebDat.german}})
             v-list-tile-sub-title {{stadien[state.neuerStatus]}} (geändert am {{state.date.german}})
-    v-dialog(v-model="addPersonShow" max-width="400px" persistend)
-      v-card
-        v-card-title
-          h1(v-font v-primary) {{addPersonenType==='add' ? 'Neues Mitglied hinzufügen' : 'Mitglied bearbeiten'}}
-        v-card-text
-          v-form(v-model="addPersonValid" @submit.prevent)
-            formular(v-model="addPersonValue" :schema=`[
-              {
-                name: 'personID',
-                type: 'autocomplete',
-                rule: 'required',
-                'prepend-icon': 'person',
-                items: (addPersonenType==='add'?allPersonen:((data.personen||[]).filter(v=>(v.currentStatus>0)).map(v=>v.person))).map(pers=>({value: pers.personID, text: pers.vorname + ' ' + pers.nachname + ' (' + pers.gebDat.german + ')'})),
-                label: 'Person'
-              },
-              {
-                name: 'status',
-                rule: 'required',
-                type: 'select',
-                items: stadien.map((besch, id)=>({value: id, text: besch})),
-                label: 'Neuer Status'
-              },
-              {
-                name: 'date',
-                type: 'date',
-                label: 'Datum des Updates',
-                rule: 'required',
-                required: true
-              }
-            ]`)
-        v-card-actions
-          v-spacer
-          v-btn(flat @click="addPersonShow=false") Abbrechen
-          v-btn(color="primary" :disabled="!addPersonValid" @click="addPersonSave") Speichern
+    ec-form-edit-ak(ref="formEditAk")
 </template>
 <script lang="ts">
 import { Component, Vue, Prop, Watch } from 'vue-property-decorator';
@@ -63,7 +31,6 @@ import { genReport } from '@/report'
 
 @Component({})
 export default class EcRootIndex extends Vue {
-
   private get config() {
     return  {
       sheet: [
@@ -71,19 +38,19 @@ export default class EcRootIndex extends Vue {
           id: 'ak_m_add',
           icon: 'person_add',
           label: 'AK Mitglied hinzufügen',
-          click: () => {this.edit('add'); }
+          click: () => {(<any>this.$refs.formEditAk).edit('add'); }
         },
         {
           id: 'ak_m_update',
           icon: 'edit',
           label: 'AK Mitglied updaten',
-          click: () => {this.edit('edit'); }
+          click: () => {(<any>this.$refs.formEditAk).edit('edit'); }
         },
         {
           id: 'ak_m_del',
           icon: 'delete',
           label: 'AK Mitglied entfernen',
-          click: () => {this.edit('delete'); }
+          click: () => {(<any>this.$refs.formEditAk).edit('delete'); }
         },
         {
           id: 'ak_rep_current',
@@ -107,16 +74,6 @@ export default class EcRootIndex extends Vue {
     };
   }
   public static meta = {};
-
-  private personenData = [];
-
-  private addPersonShow = false;
-  private addPersonValid = false;
-  private addPersonValue: any = {};
-  private addPersonenType: 'add'|'edit'|'delete'|'' = '';
-
-  private allPersonen: any = [];
-
 
   private showAll = false;
 
@@ -142,85 +99,6 @@ export default class EcRootIndex extends Vue {
     });
   }
 
-  private addPersonSave() {
-    this.addPersonShow = false;
-
-    this.$apolloClient.mutate({
-      mutation:  gql`
-        mutation(
-          $personID: Int!,
-          $akID: Int!,
-          $date: String!,
-          $status: Int!,
-          $authToken: String!
-        ) {
-          updateAKStatus(
-            personID:  $personID,
-            akID: $akID,
-            date: $date,
-            status: $status,
-            authToken: $authToken
-          )
-        }
-      `,
-      variables: {...this.addPersonValue,  akID: this.$route.params.id, authToken: this.$authToken}
-    }).then(()=>{
-      this.$notifikation('Neuer Eintrag im AK', `Du hast erfolgreich einen neuen Eintrag im AK angelegt`);
-    }).catch((err: any) => {
-      this.$dialog.error({
-        text: err.message,
-        title: 'Speichern fehlgeschlagen!'
-      });
-    });
-
-  }
-
-  private edit(type: 'add'|'edit'|'delete') {
-    this.addPersonenType = type;
-    this.addPersonValue = {};
-
-    if (type === 'delete') {
-      this.addPersonValue = {
-        status: 0
-      };
-    }
-
-    if (type === 'add' && this.allPersonen) {
-      this.getPersonen();
-    }
-
-    this.addPersonShow = true;
-  }
-
-  private getPersonen() {
-    this.$apolloClient.query({
-      query: gql`
-        query($authToken: String!) {
-          personen(authToken: $authToken) {
-            personID,
-            vorname,
-            nachname,
-            gebDat {
-              german
-              input
-            }
-          }
-        }
-      `,
-      variables: {
-        authToken: this.$authToken
-      }
-    })
-    .then((res) => {
-      this.allPersonen = res.data.personen;
-    })
-    .catch((err: any) => {
-      this.$dialog.error({
-        text: err.message,
-        title: 'Laden fehlgeschlagen!'
-      });
-    });
-  }
 
   private loadData() {
     this.$apolloClient.query({
